requirements.txt
aiohttp
pyyaml
flake8>=2.2.4
flake8-docstrings
flake8-per-file-ignores
